Separate model_functions.py from motor_model.py, unittest param save/load, nonlinparams fully plumbed through calibration
