+++ test/unittests/calibration_unittest.py
@@ -9,7 +9,7 @@ from os import path
 import pytest
 import sys
 sys.path.append(path.join(path.dirname(path.dirname(path.dirname(__file__))), "calibrate"))
-from motor_model import ModelFcn, get_nonlin_paramArr, assign_nonlin_parameters, model_fcns
+from motor_model import ModelFcn, get_nonlin_paramArr, assign_nonlin_parameters, loadParams, saveParams
 
 
 class MultiplierModelFcn(ModelFcn):
@@ -105,17 +105,15 @@ class TestModelFcn:
         assert "input" in str(excinfo.value)
 
     # Test call with model_fcns[mod](results, nonlinearparams[mod])
-    def test_call_fcn_pass_nonlinparams(self, mocker):
-        print("FIRST one")
+    def test_call_fcn_pass_nonlinparams(self):
         new_model_fcns = get_mock_model_fcns()
-        mocker.patch("motor_model.model_fcns", new=new_model_fcns)
         model = ["mult", "mult_copy", "mult_10", "add_mult", "nested_mult", "10_nested_mult"]
         param_values = np.array([2.0, 3.0,  # mult, mult_copy, none from mult_10,
                                  4.0, 5.0,  # add_mult
                                  6.0, 7.0,  # nested_mult
                                  8.0        # 10_nested_mult
                                  ])
-        nonlinear_params = assign_nonlin_parameters(model, param_values)
+        nonlinear_params = assign_nonlin_parameters(model, param_values, new_model_fcns)
         expected_contributions = {
             "mult": self.results["vel_f"] * 2.0,
             "mult_copy": self.results["vel_f"] * 3.0,
@@ -126,29 +124,28 @@ class TestModelFcn:
         }
         for mod in model:
             assert np.all(new_model_fcns[mod](self.results, nonlinear_params[mod]) == expected_contributions[mod])
+            assert new_model_fcns[mod](self.results, nonlinear_params[mod]) \
+                    == pytest.approx(expected_contributions[mod])
 
 
 class TestNonlinParamsConversion:
-    def test_get_nonlin_paramArr_simple(self, mocker):
-        print("OTHERONE")
-        mocker.patch("motor_model.model_fcns", new=get_mock_model_fcns())
-        nonlin_params = get_nonlin_paramArr(["mult", "mult_copy", "mult_10"])
+    def test_get_nonlin_paramArr_simple(self):
+        nonlin_params = get_nonlin_paramArr(["mult", "mult_copy", "mult_10"], get_mock_model_fcns())
         assert all(nonlin_params == [1.0, 1.0])
 
-    def test_get_nonlin_paramArr_nested(self, mocker):
-        mocker.patch("motor_model.model_fcns", new=get_mock_model_fcns())
-        nonlin_params = get_nonlin_paramArr(["mult", "mult_copy", "mult_10", "add_mult", "nested_mult", "10_nested_mult"])
+    def test_get_nonlin_paramArr_nested(self):
+        nonlin_params = get_nonlin_paramArr(["mult", "mult_copy", "mult_10", "add_mult", "nested_mult", "10_nested_mult"],
+                                            get_mock_model_fcns())
         assert all(nonlin_params == [1.0, 1.0, 1.0, 0.0, 1.0, 1.0, 1.0])
 
-    def test_get_nonlin_dict(self, mocker):
-        mocker.patch("motor_model.model_fcns", new=get_mock_model_fcns())
+    def test_get_nonlin_dict(self):
         model = ["mult", "mult_copy", "mult_10", "add_mult", "nested_mult", "10_nested_mult"]
         param_values = np.array([2.0, 3.0,  # mult, mult_copy, none from mult_10,
                                  4.0, 5.0,  # add_mult
                                  6.0, 7.0,  # nested_mult
                                  8.0        # 10_nested_mult
                                  ])
-        nonlinear_params = assign_nonlin_parameters(model, param_values)
+        nonlinear_params = assign_nonlin_parameters(model, param_values, get_mock_model_fcns())
         assert nonlinear_params == {"mult": {"mult": 2.0},
                                     "mult_copy": {"mult": 3.0},
                                     "mult_10": {},
@@ -156,3 +153,10 @@ class TestNonlinParamsConversion:
                                     "nested_mult": {"mult": 6.0, "input": {"mult": 7.0}},
                                     "10_nested_mult": {"input": {"mult": 8.0}},
                                     }
+
+def test_params_save_load(tmp_path):
+    params = {"horse": 1000., "battery_staple": 2.3}
+    filepath = tmp_path / "motorparams"
+    saveParams(params, filepath)
+    new_params = loadParams(filepath)
+    assert new_params == params
